requirements.txt
azure-core==1.32.0
azure-cosmos==4.9.0
azure-functions==1.21.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
dnspython==2.7.0
idna==3.10
msal==1.32.0
pycparser==2.22
PyJWT==2.10.1
pymongo==4.11.3
python-dotenv==1.1.0
requests==2.32.3
six==1.17.0
typing_extensions==4.13.0
urllib3==2.3.0
